+++ epi2025/epi/epi_app/urls.py
@@ -1,6 +1,6 @@
 
 from django.urls import path
-from . import views
+from epi_app import views
 
 urlpatterns = [
     path('', views.home, name='home'),
