+++ epi2025/epi/epi_app/views/__init__.py
@@ -1,0 +1,5 @@
+from .home.home_views import *
+from .colaboradores.colaboradores_views import *
+from .epis.epis_views import *
+from .controles.controles_views import *
+from .usuarios_login.usuarios_login_views import *
